[CONSUL-543] Create exec_supported.go
